Remove base measurements
